Split InspectableStream tests to their own file
