+++ test/inspectable.stream.mjs
@@ -1,6 +1,4 @@
 import stream from 'node:stream'
-import { describe, it } from 'node:test'
-import assert from 'node:assert'
 
 export class InspectableStream extends stream.Duplex {
   #writeBuffer = Buffer.alloc(0)
@@ -40,31 +38,3 @@ export class InspectableStream extends stream.Duplex {
     return this.#writeBuffer.length
   }
 }
-
-describe('InspectableStream', () => {
-  it('writes can be extracted', () => {
-    // Given
-    const stream = new InspectableStream()
-    const expected = 'Hello world!'
-
-    // When
-    stream.write(expected)
-
-    // Then
-    const actual = stream.extract().toString()
-    assert(actual === expected)
-  })
-
-  it('reads can be injected', () => {
-    // Given
-    const stream = new InspectableStream()
-    const expected = 'Hello world!'
-
-    // When
-    stream.inject(expected)
-
-    // Then
-    const actual = stream.read().toString()
-    assert(actual === expected)
-  })
-})
